+++ src/tools/index.js
@@ -1,5 +1,7 @@
 
 export * from "./CursorInfoTool";
+export * from "./PanDraggingTool";
 export * from "./SliceScrollingTool";
 export * from "./Tool";
-export * from "./VoiWindowingTool";
+export * from "./VoiWindowDraggingTool";
+export * from "./ZoomDraggingTool";